@@ -13,10 +13,10 @@
     "dotenv": "^16.0.3",
     "jwt-decode": "^2.2.0",
     "qs": "^6.11.0",
-    "react": "^16.13.1",
+    "react": "^18.2.0",
     "react-bootstrap": "^1.0.1",
     "react-datepicker": "^4.8.0",
-    "react-dom": "^16.13.1",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.2.1",
     "react-scripts": "^5.0.1"
   },
